修改整合后bug，增加参数检查

+++ fzutopic/src/main/java/com/fzutopic/controller/LikesController.java
@@ -2,8 +2,12 @@ package com.fzutopic.controller;
 
 
 import com.fzutopic.annotation.UserLoginToken;
+import com.fzutopic.dao.CommentlikesDao;
+import com.fzutopic.dao.TopiclikesDao;
 import com.fzutopic.model.*;
+import com.fzutopic.service.CommentServiceImpl;
 import com.fzutopic.service.LikesServiceImpl;
+import com.fzutopic.service.ReplyServiceImpl;
 import com.fzutopic.service.TopicServiceImpl;
 import com.fzutopic.utils.TokenUtil;
 import lombok.extern.slf4j.Slf4j;
@@ -26,19 +30,52 @@ public class LikesController {
     @Resource(name = "topicServiceImpl")
     TopicServiceImpl topicService;
 
+    @Resource(name = "replyServiceImpl")
+    ReplyServiceImpl replyService;
+
+    @Resource(name = "commentServiceImpl")
+    CommentServiceImpl commentService;
+
+    @Resource
+    TopiclikesDao topiclikesDao;
+
+    @Resource
+    CommentlikesDao commentlikesDao;
+
+    //用户首次对话题点赞（踩）
+    @UserLoginToken
+    @PostMapping("/user/topic/topiclikes")
+    public AjaxResponse postToplicLikes(@RequestBody Topiclikes topiclikes, HttpServletRequest httpServletRequest) {
+        String userid = TokenUtil.getUserIdByRequest(httpServletRequest);
+        String topicid = topiclikes.getTopicid();
+        topiclikes.setUserid(userid);
+        if (topiclikesDao.selectByPrimaryKey(topiclikes) != null)
+            return AjaxResponse.error(500, "用户：" + userid + "非首次对话题：" + topicid + "点赞（踩）");
+        int status = topiclikes.getLikedstatus();
+        AjaxResponse ajaxResponse = topicService.insertLikesById(topicid, status);
+        int code = ajaxResponse.getCode();
+        if (code != 200) return ajaxResponse;
+        likesService.postTopicLikes(topiclikes);
+        return AjaxResponse.success("话题点赞（踩）成功");
+    }
+
     //用户取消对话题的点赞或踩
     @UserLoginToken
     @DeleteMapping("/user/topic/topiclikes")
     public AjaxResponse deleteTopicLikes(@RequestParam(name = "topicid") String id, HttpServletRequest httpServletRequest) {
+        String userid = TokenUtil.getUserIdByRequest(httpServletRequest);
         TopiclikesKey key = new TopiclikesKey();
         key.setTopicid(id);
-        String userid = TokenUtil.getUserIdByRequest(httpServletRequest);
         key.setUserid(userid);
-        //更新话题总的点赞（踩）数
-
-        //boolean success=topicService.updateLikesById()
+        Topiclikes topiclikes = topiclikesDao.selectByPrimaryKey(key);
+        if (topiclikes == null)
+            return AjaxResponse.error(500, "不存在用户：" + userid + "对话题：" + id + "的点赞（踩）记录）");
+        int status = topiclikes.getLikedstatus();
+        AjaxResponse ajaxResponse = topicService.deleteLikesById(id, status);
+        int code = ajaxResponse.getCode();
+        if (code != 200) return ajaxResponse;
         likesService.deleteTopiclikes(key);
-        return AjaxResponse.success(id);
+        return AjaxResponse.success("取消话题点赞（踩）成功");
     }
 
     //用户修改对话题的态度
@@ -46,60 +83,118 @@ public class LikesController {
     @PutMapping("/user/topic/topiclikes")
     public AjaxResponse updateTopicLikes(@RequestParam(name = "topicid") String id, @RequestParam(name = "likes") int likes,
                                          HttpServletRequest httpServletRequest) {
-
-        Topiclikes topiclikes = new Topiclikes();
         String userid = TokenUtil.getUserIdByRequest(httpServletRequest);
-        topiclikes.setUserid(userid);
-        topiclikes.setTopicid(id);
-        topiclikes.setLikedstatus(likes);
-        likesService.updateTopicLikes(topiclikes);
-        return AjaxResponse.success(id);
-    }
-
-    //用户首次对话题点赞（踩）
-    @UserLoginToken
-    @PostMapping("/user/topic/topiclikes")
-    public Topiclikes postToplicLikes(@RequestBody Topiclikes topiclikes, HttpServletRequest httpServletRequest) {
-        String userid = TokenUtil.getUserIdByRequest(httpServletRequest);
-        topiclikes.setUserid(userid);
-        likesService.postTopicLikes(topiclikes);
-        return topiclikes;
-    }
-
-    //用户取消对评论（回复）的赞（踩）
-    @UserLoginToken
-    @DeleteMapping("/user/topic/comment/commentlikes")
-    public AjaxResponse deleteCommentLikes(@RequestParam(name = "itemid") String itemid, HttpServletRequest httpServletRequest) {
-        CommentlikesKey key = new CommentlikesKey();
-        key.setItemid(itemid);
-        String userid = TokenUtil.getUserIdByRequest(httpServletRequest);
+        TopiclikesKey key = new TopiclikesKey();
+        key.setTopicid(id);
         key.setUserid(userid);
-        likesService.deleteCommentlikes(key);
-        return AjaxResponse.success(itemid);
-    }
-
-    //用户修改对评论（回复）的态度
-    @UserLoginToken
-    @PutMapping("/user/topic/comment/commentlikes")
-    public AjaxResponse updateCommentLikes(@RequestParam(name = "itemid") String itemid,
-                                           @RequestParam(name = "likes") int likes, HttpServletRequest httpServletRequest) {
-        Commentlikes commentlikes = new Commentlikes();
-        String userid = TokenUtil.getUserIdByRequest(httpServletRequest);
-        commentlikes.setUserid(userid);
-        commentlikes.setItemid(itemid);
-        commentlikes.setLikedstatus(likes);
-        likesService.updateCommentLikes(commentlikes);
-        return AjaxResponse.success(itemid);
+        Topiclikes topiclikes = topiclikesDao.selectByPrimaryKey(key);
+        if (topiclikes == null)
+            return AjaxResponse.error(500, "不存在用户：" + userid + "对话题：" + id + "的点赞（踩）记录）");
+        if (likes != 1 && likes != 0)
+            return AjaxResponse.error(500, "likes参数值非法要求为0或1");
+        if (likes == topiclikes.getLikedstatus())
+            return AjaxResponse.error(500, "用户：" + userid + "未对话题：" + id + "的点赞（踩）作出修改");
+        topiclikes.setLikedstatus(likes);
+        AjaxResponse ajaxResponse = topicService.updateLikesById(id, likes);
+        int code = ajaxResponse.getCode();
+        if (code != 200) return ajaxResponse;
+        likesService.updateTopicLikes(topiclikes);
+        return AjaxResponse.success("成功修改点赞（踩）为踩（点赞）");
     }
 
     //用户首次对评论（回复）点赞（踩）
     @UserLoginToken
     @PostMapping("/user/topic/comment/commentlikes")
-    public Commentlikes postCommentLikes(@RequestBody Commentlikes commentlikes, HttpServletRequest httpServletRequest) {
+    public AjaxResponse postCommentLikes(@RequestBody Commentlikes commentlikes,
+                                         @RequestParam(name = "sort") int sort, HttpServletRequest httpServletRequest) {
+        if (sort != 0 && sort != 1)
+            return AjaxResponse.error(500, "sorts参数值非法，要求是1（评论）或0（回复）");
+
         String userid = TokenUtil.getUserIdByRequest(httpServletRequest);
+        String itemid = commentlikes.getItemid();
+        int likes = commentlikes.getLikedstatus();
         commentlikes.setUserid(userid);
+        if (commentlikesDao.selectByPrimaryKey(commentlikes) != null)
+            return AjaxResponse.error(500, "用户：" + userid + "非首次对此评论（回复）点赞（踩）");
+        if (likes != 1 && likes != 0)
+            return AjaxResponse.error(500, "likes参数值非法，要求为0或1");
+
+        AjaxResponse ajaxResponse;
+        if (sort == 1)
+            //表示是对评论点赞（踩）
+            ajaxResponse = commentService.insertLikesById(itemid, likes);
+        else
+            //表示是对评论的回复点赞（踩）
+            ajaxResponse = replyService.insertLikesById(itemid, likes);
+        int code = ajaxResponse.getCode();
+        if (code != 200) return ajaxResponse;
         likesService.postCommentLikes(commentlikes);
-        return commentlikes;
+        return AjaxResponse.success("成功点赞（踩）");
+    }
+
+    //用户修改对评论（回复）的态度
+    @UserLoginToken
+    @PutMapping("/user/topic/comment/commentlikes")
+    public AjaxResponse updateCommentLikes(@RequestParam(name = "itemid") String itemid, @RequestParam(name = "likes") int likes,
+                                           @RequestParam(name = "sort") int sort, HttpServletRequest httpServletRequest) {
+        if (sort != 0 && sort != 1)
+            return AjaxResponse.error(500, "sorts参数值非法，要求是1（评论）或0（回复）");
+
+        String userid = TokenUtil.getUserIdByRequest(httpServletRequest);
+        CommentlikesKey commentlikesKey = new CommentlikesKey();
+        commentlikesKey.setItemid(itemid);
+        commentlikesKey.setUserid(userid);
+        Commentlikes commentlikes = commentlikesDao.selectByPrimaryKey(commentlikesKey);
+        if (commentlikes == null)
+            return AjaxResponse.error(500, "不存在用户：" + userid + "对评论（回复）：" + itemid + "的点赞（踩）记录）");
+        if (likes != 1 && likes != 0)
+            return AjaxResponse.error(500, "likes参数值非法，要求为0或1");
+        if (likes == commentlikes.getLikedstatus())
+            return AjaxResponse.error(500, "用户：" + userid + "未对评论（回复）：" + itemid + "的点赞（踩）作出修改");
+        commentlikes.setLikedstatus(likes);
+
+        AjaxResponse ajaxResponse;
+        if (sort == 1)
+            //表示是对评论点赞（踩）
+            ajaxResponse = commentService.updateLikesById(itemid, likes);
+        else
+            //表示是对评论的回复点赞（踩）
+            ajaxResponse = replyService.updateLikesById(itemid, likes);
+        int code = ajaxResponse.getCode();
+        if (code != 200) return ajaxResponse;
+        likesService.updateCommentLikes(commentlikes);
+        return AjaxResponse.success("成功修改点赞（踩）为踩（点赞）");
+    }
+
+    //用户取消对评论（回复）的赞（踩）
+    @UserLoginToken
+    @DeleteMapping("/user/topic/comment/commentlikes")
+    public AjaxResponse deleteCommentLikes(@RequestParam(name = "itemid") String itemid,
+                                           @RequestParam(name = "sort") int sort, HttpServletRequest httpServletRequest) {
+        if (sort != 0 && sort != 1)
+            return AjaxResponse.error(500, "sorts参数值非法，要求是1（评论）或0（回复）");
+
+
+        String userid = TokenUtil.getUserIdByRequest(httpServletRequest);
+        CommentlikesKey commentlikesKey = new CommentlikesKey();
+        commentlikesKey.setItemid(itemid);
+        commentlikesKey.setUserid(userid);
+        Commentlikes commentlikes = commentlikesDao.selectByPrimaryKey(commentlikesKey);
+        if (commentlikes == null)
+            return AjaxResponse.error(500, "不存在用户：" + userid + "对评论（回复）：" + itemid + "的点赞（踩）记录）");
+
+        int likes=commentlikes.getLikedstatus();
+        AjaxResponse ajaxResponse;
+        if (sort == 1)
+            //表示是对评论点赞（踩）
+            ajaxResponse = commentService.deleteLikesById(itemid, likes);
+        else
+            //表示是对评论的回复点赞（踩）
+            ajaxResponse = replyService.deleteLikesById(itemid, likes);
+        int code = ajaxResponse.getCode();
+        if (code != 200) return ajaxResponse;
+        likesService.deleteCommentlikes(commentlikesKey);
+        return AjaxResponse.success("取消点赞（踩）成功");
     }
 
 }
